# openai-chatkit-starter-app/openapi.yaml
openapi: 3.1.0
info:
  title: Mnemonic MCP Gateway
  version: 0.1.0
servers:
  - url: https://YOUR_PUBLIC_DOMAIN_OR_TUNNEL
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - apiKeyAuth: []
paths:
  /combinations/generate:
    post:
      operationId: generateCombinations
      summary: Generate initial-letter word combinations
      description: >
        Provide Hangul letters directly, a list of seed words (from which initials are derived),
        or a target string. Returns ranked combinations from the offline lexicon.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                letters:
                  type: array
                  items: { type: string, minLength: 1, maxLength: 1 }
                  example: ["결", "근", "신", "상"]
                words:
                  type: array
                  items: { type: string }
                  example: ["결근", "신상"]
                target:
                  type: string
                  example: "결근신상"
                beam_width: { type: integer, default: 64, minimum: 1, maximum: 4096 }
                max_candidates: { type: integer, default: 20, minimum: 1, maximum: 100 }
                keep_order: { type: boolean, default: true }
                bag_mode: { type: boolean }
                include_trace: { type: boolean, default: false }
      responses:
        "200":
          description: Ranked combinations
          content:
            application/json:
              schema:
                type: object
                properties:
                  initials:
                    type: array
                    items: { type: string }
                  candidates:
                    type: array
                    items:
                      type: object
                      properties:
                        combo: { type: string }
                        words:
                          type: array
                          items: { type: string }
                        word_sources:
                          type: array
                          items:
                            type: array
                            items: { type: string }
                        word_scores:
                          type: array
                          items: { type: number }
                        coverage:
                          type: array
                          items: { type: string }
                        mode: { type: string, enum: [sequence, bag] }
                        score: { type: number }
                  trace:
                    type: array
                    items: { type: object }
        "400": { description: Bad request }
        "401": { description: Unauthorized }
        "503": { description: mnemo_mcp unavailable }

  /mnemonics/generate:
    post:
      operationId: legacyGenerateMnemonics
      summary: (Deprecated) Alias of /combinations/generate
      deprecated: true
      requestBody:
        $ref: '#/paths/~1combinations~1generate/post/requestBody'
      responses:
        $ref: '#/paths/~1combinations~1generate/post/responses'

  /lexicon/validate:
    post:
      operationId: validateWords
      summary: Bulk-check words against the offline dictionaries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [words]
              properties:
                words:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items: { type: string }
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        word: { type: string }
                        in_dict: { type: boolean }
                        has_prefix: { type: boolean }
                        sources:
                          type: array
                          items: { type: string }
                        score: { type: number }
        "401": { description: Unauthorized }
        "503": { description: trie not available }